requirements.txt
Django==1.10.2
django-uuslug==1.1.8
Pillow==3.4.2
psycopg2==2.6.2
python-slugify==1.2.1
Unidecode==0.4.19
